// packages/workflow/src/components/workflow/types.ts
export type ValueSelector = string[] // [nodeId, key | obj key path]
// ...
export enum VarType {
  string = "string",
  number = "number",
  boolean = "boolean",
  object = "object",
  array = "array",
  arrayString = "array[string]",
  arrayNumber = "array[number]",
  arrayObject = "array[object]",
  arrayFile = "array[file]",
  any = "any",
}

export type Var = {
  variable: string
  type: VarType
  children?: Var[] // if type is obj, has the children struct
  isParagraph?: boolean
  isSelect?: boolean
  options?: string[]
  required?: boolean
}

export type NodeOutPutVar = {
  nodeId: string
  title: string
  vars: Var[]
  isStartNode?: boolean
}